make bundle.bundle_export_csv() sort the data by id
For now, we're storing some bundle exports in git on the filesystem.
We want new commits to be added to the csv's, but for disk space saving
purposes, not just clobber and rewrite the entire file.  PostgreSQL
reorders results willy-nilly all the time, which would make the csv
changes look not additive, but like a bunch of changes.  So if we
sort by id, git will just see the new lines, aka all commit diffs are
just green.