Starting new approach to bundle remote push/pull
This version uses the postgres_fdw foreign data wrapper to mount remote
databases via a normal postgresql connection.  It uses IMPORT FOREIGN SCHEMA to
import the bundle schema, and then provides various comparison functions for
push, pull and merge.

It's better than the previous version that tried to use the REST endpoint,
because it just piggy-backs on postgresql's already-robust authentication and
server-to-server communication paradigms.  It's worse because it means you have
to expose port 5432 to the world.  However, with private vpn's and such, we
should be able to find decent ways around this.