flat types: endpoint fixes.  UI mostly working after import of new bundle format
TODO:
- The code in endpoint/002-helpers.sql is stuff that I had to add beyond
just the data model and request handlers to get things working.  Lots of
suspect code in there to audit and maybe refactor.

- The old meta.field_id cast to text excluded the pk_column_name attribute.
The reasoning was that it can be looked up.  But it fouls up the logic everywhere,
and when you cast a field_id to text, it has to be looked up, which is slow,
pedantic and stupid.  Stop.

- The code in endpoint/003-utils-suspect.sql.offline is code that was removed
from the endpoint server and has not been missed (so far).  If something is
missing or broken, look there

- Next time, don't do an endpoint refactor in the middle of a meta refactor.