versioned ids
A new type of primary key, for use with bundles, designed to allow
multiple versions of the same bundle to be checked out simultaneously.
It is a composit type that contains two variables, id and commit_id.

id is a unique "identity" id, that serves the same function as normal
primary keys.  It is instantiated at row creation time and should never
be changed.

commit_id is a reference to the commit that this row is a member of.

Bundle.commit() will need to be modified to update these on commit,
and the whole logic of co-existence of checkouts needs to be worked out.