bundle: Huge optimization of head_db_stage
Every time the stage was loaded (it's refreshed on almost every action in the
IDE), head_db_stage was comparing the entire database to all bundles using
meta.field_id_literal_value, regardless of putting a `where bundle_id=x` in the
query.  Query planner wasn't pushing bundle_id down into it, so it was doing a
query for every column of every row in bundle in every bundle.

Found a pattern I really like to solve it:  Make a function for each view that
returns SETOF {view_name} which is a type.  Then you can do way smarter stuff
passing in filters and don't have to rely on the query planner so much.  Much
more exploring to do here.