bundle: audit use of "set local search_path", which was breaking tests
Set local search_path is local to the TRANSACTION not the funciton, so if you set it,
it fouls up the search_path of anything else in the transaction.  So, setting it to ''
is necessary, to work around information_schema localizing identifiers, but it needs
to be set back, even if we're doing set local search_path.  So might as well just use
search_path.