bundle: audit nulls in blob
Sometimes fields have NULL values that need to be eplicitly set on checkout.  I
did an audit of how NULLs are handled in blob, and it was way wrong.  Null
values were simply being excluded from rowset_row_field, which means that a
nullable column with a default value would get their *default value* restored
rather than an explicit NULL.  Also it causes mayhem when trying to detect
schema changes (especially added columns) in bundle.

Not sure if there ever was a coherent policy, but Now it works like this:

The NULL value is not stored in blob.  Here's why:  We could simply store a row
for NULL values in blob, except that NULL != NULL, so when we join into blob
on, say, rowset_row_field.value_hash = blob.hash, when both are NULL (as is the
case when we're storing a field's NULL value), it doesn't match.  We could use
IS NOT DISTINCT FROM, but it doesn't use indexes, so from limited testing, it
is 490x slower, which makes everything involving blob crazy slow.

SO, NULL is not stored in blob, and as such, always left join into against
blob, so that if it's null, the hash and value will both also still be null.

This really isn't ideal, because we you left join into blob, we can't
differentiate between the meanings of NULL, it's eithe "this is a stored NULL"
or "this hash does not have a blob." The latter should never happen, but...
it's a potential leak.  Can't think of anything better.  A function blob()
would be crazy slow.  IS NOT DISTINCT FROM is crazy slow.  Hmm.